Annotated Loaded Video Classification MATLAB Code
This script loads in a video file to MATLAB and then each frame of this video is displayed to the user and then passed through the now trained model, returning a label and a score reflecting the models confidence in the returned label. The score will be will be a value between 1 and 0 - a score of 1 reflects the the model is certain of the returned label, a score of 0 reflects the lowest possible confidence in returned label.